requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-dotenv==1.0.0
Werkzeug==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
pymysql==1.0.2

